Updated tahrir model to accept BitInt values for Person.id

+++ tahrir/model.py
@@ -5,7 +5,7 @@ import simplejson
 
 from sqlalchemy import (
     Column,
-    Integer,
+    BigInteger,
     DateTime,
     Unicode,
     ForeignKey,
@@ -29,7 +29,7 @@ DeclarativeBase.query = DBSession.query_property()
 
 class Issuer(DeclarativeBase):
     __tablename__ = 'issuers'
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     origin = Column(Unicode(128), nullable=False)
     name = Column(Unicode(128), nullable=False, unique=True)
     org = Column(Unicode(128), nullable=False)
@@ -60,7 +60,7 @@ class Badge(DeclarativeBase):
     description = Column(Unicode(128), nullable=False)
     criteria = Column(Unicode(128), nullable=False)
     assertions = relationship("Assertion", backref="badge")
-    issuer_id = Column(Integer, ForeignKey('issuers.id'), nullable=False)
+    issuer_id = Column(BigInteger, ForeignKey('issuers.id'), nullable=False)
 
     def __unicode__(self):
         return self.name
@@ -82,7 +82,7 @@ class Badge(DeclarativeBase):
 
 class Person(DeclarativeBase):
     __tablename__ = 'persons'
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     email = Column(Unicode(128), nullable=False, unique=True)
     assertions = relationship("Assertion", backref="person")
 
@@ -121,7 +121,7 @@ class Assertion(DeclarativeBase):
     id = Column(Unicode(128), primary_key=True, unique=True,
                 default=assertion_id_default)
     badge_id = Column(Unicode(128), ForeignKey('badges.id'), nullable=False)
-    person_id = Column(Integer, ForeignKey('persons.id'), nullable=False)
+    person_id = Column(BigInteger, ForeignKey('persons.id'), nullable=False)
     salt = Column(Unicode(128), nullable=False, default=salt_default)
     issued_on = Column(DateTime)
